requirements.txt
python-dotenv==1.0.0
discord.py==2.3.2
requests==2.31.0
simplejson==3.19.1
dash==2.14.0